+++ unix/frozen-modules/sim_quickstart.py
@@ -21,12 +21,8 @@ if '-s' in sys.argv:
 if '--addr' in sys.argv:
     # Address Explorer
     from main import numpad
-    numpad.inject('4')
-    numpad.inject('y')
-    numpad.inject('4')
     numpad.inject('8')
     numpad.inject('8')
-    numpad.inject('y')
     numpad.inject('y')
     numpad.inject('4')      # skips warning!